% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{as_flextable.regimen_costs}
\alias{as_flextable.regimen_costs}
\title{Convert regimen_costs to flextable}
\usage{
\method{as_flextable}{regimen_costs}(x, ...)
}
\arguments{
\item{x}{A \code{regimen_costs} object.}

\item{...}{Additional arguments (unused).}
}
\value{
A flextable object.
}
\description{
Convert regimen_costs to flextable
}
