% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_med_costs.R
\name{calculate_single_regimen}
\alias{calculate_single_regimen}
\title{Internal helper: calculate costs for a single regimen}
\usage{
calculate_single_regimen(
  regimen,
  model_cycle_length = 7,
  n_cycles = 52,
  cycle_times = NULL,
  subcycle_precision = FALSE,
  daily_detail = FALSE,
  simple_daily_cost = FALSE
)
}
\arguments{
\item{regimen}{A \code{med_regimen} object.}

\item{model_cycle_length}{Numeric model cycle length in days.}

\item{n_cycles}{Integer number of model cycles.}

\item{cycle_times}{Numeric vector of cycle start times.}

\item{subcycle_precision}{Logical.}

\item{daily_detail}{Logical.}

\item{simple_daily_cost}{Logical.}
}
\value{
A data.frame.
}
\description{
Internal helper: calculate costs for a single regimen
}
\keyword{internal}
