% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_single.R
\name{cost_iv_vec}
\alias{cost_iv_vec}
\title{Vectorized IV drug cost}
\usage{
cost_iv_vec(dose_mg, vial_sizes, vial_costs, wastage_threshold = 0)
}
\arguments{
\item{dose_mg}{Numeric vector of doses in mg.}

\item{vial_sizes}{Numeric vector of available vial sizes (mg).}

\item{vial_costs}{Numeric vector of corresponding vial costs.}

\item{wastage_threshold}{Fraction 0-1.}
}
\value{
Numeric vector of drug costs.
}
\description{
Accepts a vector of doses and returns a vector of costs.
}
\keyword{internal}
