% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_single.R
\name{cost_oral_vec}
\alias{cost_oral_vec}
\title{Vectorized oral drug cost}
\usage{
cost_oral_vec(
  dose_mg,
  available_strengths,
  cost_per_tablet = NULL,
  cost_per_strength = NULL
)
}
\arguments{
\item{dose_mg}{Numeric vector of doses in mg.}

\item{available_strengths}{Numeric vector of tablet strengths (mg).}

\item{cost_per_tablet}{Numeric scalar or NULL.}

\item{cost_per_strength}{Numeric vector or NULL.}
}
\value{
Numeric vector of drug costs.
}
\description{
Accepts a vector of doses and returns a vector of costs.
}
\keyword{internal}
