% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_single.R
\name{cost_single_dose}
\alias{cost_single_dose}
\title{Cost of a single dose (dispatch by route)}
\usage{
cost_single_dose(dose_mg, reg)
}
\arguments{
\item{dose_mg}{Numeric. Dose in mg.}

\item{reg}{A med_regimen object (for route, vial/tablet info).}
}
\value{
Numeric scalar: drug cost for this dose.
}
\description{
Cost of a single dose (dispatch by route)
}
\keyword{internal}
