% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_single.R
\name{cost_single_dose_vec}
\alias{cost_single_dose_vec}
\title{Vectorized cost dispatcher}
\usage{
cost_single_dose_vec(dose_mg, reg)
}
\arguments{
\item{dose_mg}{Numeric vector of doses in mg.}

\item{reg}{A med_regimen object.}
}
\value{
Numeric vector of drug costs.
}
\description{
Vectorized cost dispatcher
}
\keyword{internal}
