% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{theme_regimen}
\alias{theme_regimen}
\title{Apply standard formatting theme to a regimen flextable}
\usage{
theme_regimen(ft, cost_cols = NULL, int_cols = NULL, mg_cols = NULL)
}
\arguments{
\item{ft}{A flextable object.}

\item{cost_cols}{Column names to format as currency.}

\item{int_cols}{Column names to format as integers.}

\item{mg_cols}{Column names to format with 2 decimal places.}
}
\value{
A formatted flextable.
}
\description{
Apply standard formatting theme to a regimen flextable
}
\keyword{internal}
