% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_regimen.R
\docType{data}
\name{VIAL_ROUTES}
\alias{VIAL_ROUTES}
\title{Routes that use vial-based costing (greedy packing, wastage threshold)}
\format{
An object of class \code{character} of length 3.
}
\usage{
VIAL_ROUTES
}
\description{
Routes that use vial-based costing (greedy packing, wastage threshold)
}
\keyword{internal}
