% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_single.R
\name{cost_single_admin_drug}
\alias{cost_single_admin_drug}
\title{Cost of a single administration (dispatch by route)}
\usage{
cost_single_admin_drug(reg)
}
\arguments{
\item{reg}{A med_regimen object.}
}
\value{
Numeric scalar: drug cost only (no admin cost) of one admin.
}
\description{
Cost of a single administration (dispatch by route)
}
\keyword{internal}
